Add log + validators file

+++ index.js
@@ -1,6 +1,7 @@
 const chokidar = require('chokidar');
 const fs = require('fs');
 const moment = require('moment');
+const isValidEntity = require('/Dev/ClinicInterface/validators');
 
 const fields = ['Num_Avi', 'Emer_Dept_hmo_GK', 'Status_Clinic', 'Status_Pharmacy', 'Status_Family_Doctor', 'Status_Nurse', 'Status_Pediatrician', 'Status_Gynecologist']
 const dirPath = 'ex-files'
@@ -9,14 +10,28 @@ const watcher = chokidar.watch(dirPath, { persistent: true });
 watcher.on('add', newFileListener);
 
 function newFileListener(path) {
+    console.log(`New file was added: ${path}`)
     let entityArr = readNewFile(path);
+
+    let invalidEntities = entityArr.filter(entity => !isValidEntity(entity));
+    console.log(`Found ${invalidEntities.length} invalid entities:`);
+    invalidEntities.forEach((entity, index) => console.log(`Entity number ${index + 1} is invalid`));
+
+    entityArr = entityArr.filter(entity => isValidEntity(entity));
+    if(entityArr === []){
+         console.log("No valid entities.");
+         return;
+    }
+
     let jsonArr = entityArr.map(entity => buildJson(entity));
     sendToDP(jsonArr);
+
     fs.unlinkSync(path); // delete the file 
 }  
 
 function readNewFile(path) {
     let file =  fs.readFileSync(path, 'utf8');
+    console.log(`File data: ${file}`);
     let lines =  file.split('\r\n')
     return lines.map(line => line.split(';'));
 }
@@ -38,3 +53,4 @@ function parseToISOString(date, time){
 function sendToDP(jsonArr){
     console.log(jsonArr);
 }
+
